2021: Day 12, optimize
Part 2 turned out to be slow due to algorithmic complexity. Combined
with cloning the paths over and over this made the solution slow. Since
the order of the paths aren't interesting for the answer there's no
point keeping ordered vectors of the paths. Instead, the recursion
ensures nodes are visited properly, while just count the number of
sub-paths in each branch of the recursion.